Spreadsheets Example: Fix bug in inserting a new row
When inserting a new row (above or below an existing row), it was
causing a crash.

There was a forgotten call of the previous() method on the iterator
that was supposed to iterate backward.

It has been fixed by adding this function call inside the loop.

Task-number: QTBUG-125767
Change-Id: I6a4436bacb0348c1b6bccd929f715e36d0279754
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 33b5ba74ed99a1d19dcd757f941f7909f9d5c414)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>