QQuickTextNodeEngine::addText(): don't crash if QTextLine is invalid
When a QTextBlock has been pulled out of a QTextDocument,
lineNumberForTextPosition() returns -1, so lineForTextPosition() returns
a default-constructed QTextLine whose `eng` pointer is null.
QTextLine::textStart() needs to access `eng`, so we can't call it.

The new autotest is similar to what a syntax highlighter does.
However QTextLayout::setFormats() doesn't trigger QTextEdit::update()
by itself. QTextCursor::endEditBlock() results in QTD::contentsChanged()
being emitted, which is connected to QQuickTextEdit::q_textChanged();
so we test it both ways to ensure that it doesn't crash in either case,
but one can see the text turn green only in the case that update() is
called somehow.

Done-with: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Fixes: QTBUG-126886
Pick-to: 6.8 6.7 6.5 6.2
Change-Id: I9c1093af7002432add3b4ab856a41ce2f9af4619
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>