Modify tst_popup.qml::test_anchors() to accommodate PopupWindow popups
The test test_anchors() in tst_popup.qml checks the success of centering
a popup by querying its local position. This does not work for top level
windows and so an extra calculation has to be added using the global
co-ordinates of the popup to find its local position. There is also a
bug when centering the popup in the overlay, and the overlay is not the
parent. This bug is a result of not translating the expected results to
the co-ordinate system of the parent.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: I9605b6fd290180f777afd35573163fd3c9eff95f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>