QJSEngine: Disable tail calls for GeneratorFunction/Method
When a generator function or method is called, the engine needs to
perform some setup for the call to be performed.

For example, it needs to setup an actual generator object, provide it
with a frame for re-entering, actually execution the body of the
generator callable and then return the generator object as the result.

This necessary processing is generally handled by
`GeneratorFunction::virtualCall`.

The engine implements Proper tail calls as introduced by the ES6
specification.

For example in:

```
function foo() { return 0; }
function bar() { return foo(); }
```

The call to `foo()` in `bar()` in in a tail position and can be
optimized accordingly.

In the engine, tail calls will be optimized and implemented as jumps,
eliding the actual call indirection.

While this is generally fine, it produces incorrect results when the
elided call is to a generator function/method, as the required
setup for the generator call is never executed.

For example in:

```
function* foo() { yield 0; }
function bar() { return foo(); }
```

`foo` is a generator function that is called in tail position in `bar`.
When `bar` is called, the call to `foo` will be optimized out and its
body will be executed.
This will correctly execute the body but produce an incorrect result.
The YIELD instruction will mark the incorrect frame as a yielding frame
and no generator object will be constructed or returned, instead
producing an undefined result.

To avoid the issue, disable TCE for
`GeneratorFunction`/`GeneratorMethod` to ensure that tail calls to a
generator function or method will not elide the call to
`GeneratorFunction::virtualCall`, so that they can perform the required
setup for the execution to succeed.

Some test cases were added to `tst_QJSEngine` to test that tails calls
to a generator function or method do not produce an undefined value.

Fixes: QTBUG-109880
Change-Id: I712ab5c8fdcfde2591b83c4e951dd4723b8bfc1d
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 07652bcbdc25bd754a3b22b8ae3d0e00e05f35d8)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit d9655bd538997308b1ab07b76bca1d62b9b43fd7)
(cherry picked from commit e9e5e5ac7396594ce511a0796c95137fc7bb736b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>