semantichighlightin: Fix occasional crash
This assert is triggered in rare times when editing the file. The issue
is not reproducible deterministically but can be encountered by
constantly edit the file into an invalid file and make it valid qml file
again. At some point, the qmlfile becomes invalid while it is still
processing highlighting tokens.

Pick-to: 6.8
Change-Id: Ie31ee848c5ecc1d045eecc7797af7f4e6e837a0f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>