doc: Organize internal Qt Quick and QML docs for doxygen
The \mainpage tag should only be used in one place in an entire set of
docs. The top-level about pointer event delivery is certainly not the
index of docs for all of Qt, so use \page and allow it to be listed
along the left-side index somehow arbitrarily (alongside some others
which are still not well classified). It doesn't seem easy to organize
all the internal docs into one big tree; but organizing parts of the Qt
Quick docs that way might be reasonable. Amends
a70b9c4e7bfd4cb9116dcb676965b59118e4a221

Change-Id: I70a8f437b7c80f3ca86cbac183a39138bc3696e9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>