Material Style: update style theme when system theme is changed
The Material style did not use the system theme when it was set to
QQuickMaterialStyle::System. The problem was that QQuickMaterialStyle
does not know about system theme changes.

To make this possible, it is required to implement the
QtQuickControls2MaterialStylePlugin::updateTheme() method. This method
overrides the QQuickStylePlugin::updateTheme() virtual method, and it is
when the system theme is changed. The material style plugin has access
to the QQuickMaterialTheme. Then, the material theme will be notified
when the system theme is changed.

The material theme did not have access to the material styles as they
are attached to the QML objects in the QML engine. To address this, when
a material style's theme is set to System, that material style can
register itself to the list of the system styles that are stored in the
material theme.

When the system theme is changed, the material style plugin notifies the
material theme and the material theme iterates through the material
system styles and updates their themes.

To prevent dangling pointer issues, the material styles list (with
System theme) uses QPointer type to detect null pointers after they are
destructed. They will be removed from the list when they are touched
and are null.

The testcase creates an ApplicationWindow with value of Material.System
for Material.theme property to follow system theme changes. It then
toggles the platform theme through the TestHandler and compares the
results. The TestHandler class creates a mocked class called
MockPlatformTheme that inherits from the QPlatformTheme class and
overrides the colorScheme() and requestColorScheme() virtual functions.
The reason for overriding these methods is to simulate the platform
theme change event.

[ChangeLog][Controls][Material] If the Material.theme is set to
Material.System, the application theme changes when the system theme is
changed. This also works for the child attached styles. If its theme is
set to Material.System, regardless of its attached parent style, it will
follow the system theme changes.

Fixes: QTBUG-127169
Change-Id: I29a0c59525f342595a20a908faa85bcae6615bf4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit b45629207ee32a3d80b6ea6553e8762eec8a86da)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>