Many tests fail when a dimmer is created using PopupWindow popupType
When the dimmer is created, the item is attempted to be stacked before
the top level PopupWindow. This of course is unnecessary and results in
a failed test (QQuickItem::stackBefore throws a warning when it is
unsuccessful).

Fixes: QTBUG-127166
Change-Id: I464718a176a93008bdd6c7545e6a4ae401566d89
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit d0c358b1ee64f2a27b69ce575e625a65fd4dcc4c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>