Compiler: Properly read value type arguments when enforcing signatures
The patch that introduces the enforcing of signature types did so by
passing the contained rather than the stored metatypes for the
arguments.

These types are used to populate the registers of the function with the
arguments using the proper types by static_cast'ing the void* arguments
to the actual types.

However, for value types, the arguments were assumed to be QVariants and
were thus casted to one even though they were actually passed as the
actual argument type. This seems to have been mostly fine by accident
because of the inline storage of QVariant that lays at offset 0 in its
layout. Therefore, if the flag signalling that the value is actually
elsewhere was not set to 1 by the casting and the value fit in the
inline storage everything would still work. This is not always the case
however and can lead to crashes.

Therefore, treat value type arguments as plain non-wrapped values when
populating function arguments.

Amends 8bf5aae19b77b618f3f7a55a59e87c8a319475a8

Fixes: QTBUG-127009
Change-Id: I495bcff7631399f207d87fea698d7e921e8e4721
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 179581e16fdcb071cec40cd7189251ea9ac08d10)
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>