QQuickAnimation: Fix assert condition
In general, we expect transition to return a new job. However, returning
a nullptr is also a possibility, and in that case, having the same value
as the old one is okay.
The code actually already handled that case just fine, as deleting a
nullptr is not an issue.
Amends 17fa564a94f98486ff41030c47629bd902c48e8d

Pick-to: 6.8 6.7 6.5
Task-number: QTBUG-125571
Change-Id: Iad730cb04aa1f94249a0f2aee1e5844f5e5e01df
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>