DocumentSymbol introduce SymbolKindOf helper
This commit introduces a mapping between DomType and SymbolKind.
It was decided to proceed with the static fixed-sized array instead
of map or hash-map due to the knowledge of size beforehand, plus
very small size of the mapping collection.
This TypeSymbolRelations are also utilized to filter out DomItem-s,
which are not supported, like Dom::List, Dom::Map, etc.

Note: Due to the current Dom design, SymbolKind for MethodInfo needs
to be handled separately

Task-number: QTBUG-120002
Change-Id: Icadd89e77f5d4c71215ee10cd921c2cd915e771d
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>