Skip mouse propagation test for popup window popup types
The test test_mousePropagation() checks that the mouse does not
propagate through a popup window. This is only necessary to test for
popupItem popup types, as they exist in the same window. However, Popup
Window popup types exist in separate windows, and therefore the test
does not need to run.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: I9a3ce7afbe66dad77adbfbe93a3735e7d3d6ad0c
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>