DocumentSymbol add Detail deriving functionality
This commit introduces support for detail field of DocumentSymbol.
For Dom::Id it's the .name()
For EnumMember is implicit or explicit .value()
For QmlObject it's
- .idStr if it exists
- "root" if .idStr doesnt's exist + object is indeed root
- null otherwise
For MethodInfo it's a signature, however, because this field
is absent from the MethodInfo struct, it needs to be derived here.
Dom lacks unified support for querying signatures, esp. of "Singal"-s,
hence it's manually handled and a bit messy.
As for Binding, detail is constructed using BindingValue

Task-number: QTBUG-120002
Change-Id: I96ed605d96caaa778e09a156bed51ac4a4f09b50
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>