PopupWindow: avoid potential infinite recursion
tst_QQuickColorDialogImpl::workingInsideQQuickViewer(visible: true)
would crash with a stack overflow, when making QQuickColorDialogImpl use
PopupType::Window as it's popupType.

The patch that introduced PopupWindows (2874586cc848397ab649b49b2d39d),
used to execute the following statements in hideEvent():

   QQuickPopupPrivate::get(popup)->visible = false;
   emit popup->visibleChanged();

But this was later changed in a6d0dcdb9dc5162659f84abceaf2e30775930e47,
to be:

    if (QQuickPopup *popup = d->m_popup)
        popup->setVisible(false);

Which can cause infinite recursion, due to an eventual
QWindowPrivate::setVisible(false) call, which sends a hideEvent, which
repeats back to QQuickPopupWindow::hideEvent().

Fix the issue by overriding QWindowPrivate::setVisible(), to exit early,
in case we're in a loop.

Pick-to: 6.8
Change-Id: I55665617b46587fdd243e6dced28d4b14f966515
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>