DocumentSymbol filter out some PropertyDefinitions
Because of the way Dom stores PropertyDefinitions and Bindings,
the following expressions like `property var a: 4;` are stored
inside DOM (QmlObject to be precise) as first: propertyDefinition of `a`,
second as binding `a:4;` Hence leading to "duplicates" in the Outline view.
It was decided that in such cases "representational "priority
should be given to Binding, hence it's necessary
to filter out PropertyDefinition.
This is done using the ColonTokenRegion of PropertyDefinition from the FileLocations

Note: this could have been done later, during the rearrangement of
DocumentSymbol tree, however, to do that during the "first" visit "should be"
more computationally efficient.

Task-number: QTBUG-120002
Change-Id: I8cef8d3205c8559556937f4ccc56924ead0434f3
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>