(List|Grid)ViewAttached: Narrow down the view type
The classes tried to reuse as much functionality as possible by putting
it into the common base class, QQuickItemViewAttached. That however
meant that the view property could only have the type QQuickItemView,
confusing tooling.

This change moves the property to the derived classes, and changes the
type to the derived type (matching what is already documented).
This is safe (API stability wise) because QQuickItemViewAttached does
not exist as its own type in QML, and the C++ type exists only in a
private header.

Fixes: QTBUG-124412
Pick-to: 6.8
Change-Id: Ic613d2afce311d974072913a110a601ce2ee0251
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>