DocumentSymbol reorganizeForOutlineView
Unfortunately, there is discrepancy between the
semantic, structural and "DOM" representation of Items of the document,
therefore some reorganization needs to be done.

Namely, DocumentSymbols, representing DomItem::Id-s and
DomItem::EnumDecl(along with its children),
which are present at the level of QmlComponent need to be
"pushed down" the tree and "readopted" to the structurally relevant
parent Symbols. Same procedure applies to the inline components.
In the end the top Symbol should be the one representing root QmlObject
of the main QmlComponent (this one is coming purely from the assumption that
this will be more user-friendly)

Details
reorganizeForOutlineView expects a SymbolsList.
It requires explicitly that the first element is a Symbol representing QmlFile.*
First, reorganization within each QmlComponentm is taking place,
then each inline Component is "pushed down".
Order in the "children"-s list is not important,
because Outline client is taking care of it and sort Symbols with respect to Ranges.
Ranges should not be touched or readjusted though,
because they respect structural hierarchy.

* This is a bit of a short-cut to simplify the reorganization,
by being clear about the expectations:
QmlFile contains Symbols representing QmlComponents as children,
where the first QmlComponent is the main one, containing at least one
symbol representing QmlObject.
Alternatively it would require to do generic lookups first, to identify
QmlComponents and then do the rest.

Task-number: QTBUG-120002
Change-Id: Ica882182b025d36d38aa3fa283e8b8ecaf7c7b49
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>