tst_qmllint: Use custom qmltypes file for callJSValue
The `callJSValue` test case in `dirtyQmlCode` depends on having access
to a `QJSValue` property for the test to behave correctly.

Currently, the test uses `QQuickItem::gradient`, which is a `QJSValue`.
As the property is external to the test and may change in type, be
removed or similar issues, the test case risk breaking due to reason
unrelated to its purpose.

To avoid the issue and make the test more resilient, a custom `qmltypes`
file was added that provides a type with a `QJSValue` property.

The `qml` file used by the test was modified to use the newly provided
type and its `QJSValue` property.

The expected error message for the `callJSValue` test case was modified
to reflect the changes, in particular the change in the name of the
relevant `QJSValue` property.

Fixes: QTBUG-104597
Change-Id: I225fcd20f8fe3256c3a5dd0be843db61f7916270
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>