QQmlPropertyValidator::validateLiteralBinding: Support more numeric types
While usage of (unsigned)(long) long can result in lossy conversions, as
the QML engine only can use double and int, we already allow it at
runtime.
Teach the property validator to accept number literals when assigning to
such properties. No attempt is made to uncover assignments losing
information (as we do for int). If we really need it, qmllint warnings
are probably a better place for them than hard engine errors.

Pick-to: 6.8
Change-Id: I2329eb5a4bdde90932ee964300c7569eea2fc4e0
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>