semantichighlighting: add more property kinds
Qml editor in Qt Creator discerns property tokens by their defined
scope. It has QmlScopeObjectProperty kind used for the properties
defined in the current object. QmlRootObjectProperty used for the
properties defined in parent the objects and QmlExternalProperty type
used for the properties that are defined in the root item of the
external file.

Pick-to: 6.8
Task-number: QTBUG-126550
Change-Id: I51c624a4f4a9063e1716218e9fe70dd1234a5ec3
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>