Avoid leaking nodes in when curve renderer is destroyed
For each visual change, the curve renderer generates new scene graph
nodes in runner objects. If the renderer was destroyed before the item
was updated, the runners and their nodes would be leaked (always in
sync mode, and sometimes in async mode).

Fix by making the renderer clean up the runners on exit, and the
runners clean up any nodes that have not been inserted into the scene
graph.

Fixes: QTBUG-127519
Pick-to: 6.8
Change-Id: I9fcfe9d0b67cfa5b4c6fcef59cc5f78c579ec7ea
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>