Add tst_QQuickPinchHandler::dragVsPinch
Test the case where two touchpoints are handled by some combination
of DragHandlers and a PinchHandler on an underlying item.

The details of how this works are described in internal docs such as
src/quick/doc/src/internal/ideal-pointer-event-delivery-parallel-drags.dox
and it's easier to have an autotest that reproduces the scenarios to
verify, rather than needing to repeatedly test by hand on a touchscreen.
Anyway it wasn't covered in existing autotests until now; and it's also
good to verify the same snippet that we are describing in API docs:
src/quick/doc/snippets/pointerHandlers/pinchAndDragHandlers.qml

Change-Id: I772235c15e9a20d93769c844ae6a7f488a39c3b6
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>