qmllint: unify missing property warning
Currently, there are two different warning formulations when binding
something to a property that does not exist. Unify them so that they
don't need to be documented twice by replacing
```
Binding assigned to XXX, but no property XXX exists in the current
element.
```
with the warning
```
Property X does not exist
```

Also, from a user perspective it does not make sense to have different
warnings for "assigning an object to a property that does not exist" and
"assigning a script expression to a property that does not exist".

Pick-to: 6.7
Task-number: QTBUG-118112
Change-Id: Iec2b5717709239d4b8fa45793ee16c6e1cc39a24
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 85e47ec482072531d59c40555dea4bd5df5e63dc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>