QQuickDrag: Parse "text/uri-list" more generously
This change allows "text/uri-list" to be given as a direct string, while
still keeping the URI validity checks and fixes a regression that was
introduced in 17318c4805d089f10c25f8165dae7d1f908bd311.

Before that commit, the "text/uri-list" mimetype could be set via a
string following RFC2483.
https://www.ietf.org/rfc/rfc2483.txt

RFC2483 essentially specifies that "text/uri-list" properties are split
by CRLF with each line indicating one URI, for example the string
"file:///tmp/a.txt\r\nfile:///tmp/b.txt" would be interpreted as a list
of two local files.

And indeed, setting Drag.mimeData's "text/uri-list" to such a string
worked before the above commit.

The commit responsible for the regression introduces some sanity checks
(which are good), but also regresses the string assignment, by always
assuming that there is only one URL available and hardcoding the urls
list to just one URL:
mimeData->setUrls({url});

It was however at least possible to properly set a list of URLs from QML
by assigning a QVariantList (or a direct QByteArray).
While this is a neat convenience function, the original way of being
able to directly set RFC2483 conformant strings should be restored.

This patch fixes the regression by properly splitting the URI list again
on CRLF and it is also easy to verify that this patch does not introduce
any new regression for any behavior that was working before:
- If the uri list does contain CRLF, then before this patch it would be
  interpreted as a single URI, which is not a valid URI (the CRLF would
  need to be URI-encoded)
- If the uri-list does not contain CRLF, then the behavior is unchanged
  (a single URI is added)

A simple way to test this patch is to set:
Drag.mimeData: { "text/uri-list": "file:///tmp/a\r\nfile:///tmp/b" }

Pick-to: 6.5 6.7 6.8
Change-Id: I5fbbc1a6b92a8782d62efb94b07a1a18a33c88fb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>