Fix rendering errors with Calibri Light and curve renderer text
When selecting a base font size for the curve renderer, we chose a
rather small size, since it did not seem to make any difference.
But it turns out that this font size exhibits issues with certain
characters in certain fonts.

There is no need to keep this value low, since the size of the
font does not impact the memory usage in this case. But of course
it should not be so big that we encounter high number precision
issues instead.

A pixel size of 64 is frequently used as a base font size in Qt,
because it is typically big enough to be unhinted. It also fixes
the issue here, so we default to that instead.

[ChangeLog][Text] Fixed an issue where some fonts would exhibit
artifacts when renderered with the CurveRendering render type.

Fixes: QTBUG-126616
Change-Id: I3fe5f486ae58a6d08be90a1ebda862740ff04ba2
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 862d229666d2405e3e6904192a78f76897a4b07a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 1a0d477fb0cd88f0190aee4c6c12dcdecb8f98cd)