qmlls: fix qualified module completion
Fix the use of module qualifiers when generating snippets for all
three completion cases.
Let `import QtQuick as QQ` be an example, where the user wants to
complete a State snippet:
1. `QQ` -> insert `QQ.State { ... QQ.PropertyChanges { ... } ... }`
2. `QQ.` -> insert `State { ... QQ.PropertyChanges { ... } ... }`

The third case arises when there QtQuick is imported without qualifier:
3. `Sta` -> insert `State { ... PropertyChanges { ... } ... }`

Add tests for all three cases, and a method that creates the snippets
with the correct qualifiers. This avoids:
* getting `QQ.QQ.State ...` completions with duplicate qualifier
* getting `QQ.State ... PropertyChanges ...` completions with
  unqualified PropertyChanges

Also fix the snippet completion to avoid suggesting the wrong snippets
inside of bindings, like
`someProperty: stat` -> `someProperty: states: [...]`
for example.

Also fix the indexes in the tests to reflect the new policy for
completion on `.`.

Pick-to: 6.8 6.7
Fixes: QTBUG-127609
Change-Id: I1c70c661a15961b99157d9dc9e8de5e3320433c3
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>