qmlls: fix completion on field member expressions
Do not confuse completion requests on dots with completion requests on
identifiers.

Fix itemsFromTextLocation to return the binary expression instead of
the right hand side of the fieldmemberexpression for the location of
the dot: for example, if you try to get the DomItem for the `.` in
`object.member`, then you would previously get the DomItem for the
ScriptIdentifierExpression `object`. With this patch, it will return
the BinaryExpression `object.member`.

We usually assume that sourcelocations have inclusive ends, for example
we assume that auto-completion on `\n` in `someName\n` wants suggestions
for `someName`, even if its technically one position "outside" the
sourcelocation of `someName`. This is not true for
ScriptBinaryExpressions, where auto-completion on `.` in `someName.` should
not return suggestions for `someName`.
The same also applies to probably all binary expressions `+`, `-`, and
so on. Therefore exclude the position behind the sourcelocations when searching
for items inside ScriptBinaryExpressions in itemsFromTextLocation.

This requires adapting qmllsutils to the "new fieldmemberexpressions" as:
* ScriptIdentifierExpressions for completion inside of them, for example
  `member` in `someObject.member`. isFieldMemberAccess() returns true on
  those.
* ScriptBinaryExpressions for completion on `.`.
  isFieldMemberExpression() returns true on those.

Adapt the offsets of the tests on completion on `.`.

This breaks some tests on qualified module identifiers, so QEXPECT_FAIL
them and track their resolution via QTBUG-127609.

Pick-to: 6.8 6.7
Fixes: QTBUG-127586
Change-Id: I82726205c7401e43174cc7749372e0776063bcbb
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>