QJSEngine: Check for stack overflows on generators
Generally, the engine checks for stack overflows each time a function is
called.

In the common case, the entry point for the execution of a function is
`QV4::Moth::VME::exec`, which will perform the check as part of its
execution.
`exec` will then dispatch to the execution of the function, either by
using the JIT when enabled and available or by calling
`QV4::Moth::VME::interpret` when this is not the case.

Generators work slightly differently than a normal call, both when a
Generator Function or Method is executed and when a generator is
resumed.

In particular, the execution of a Generator Function or Method and the
execution of a resumed generator do not pass by `exec`, instead directly
calling `interpret`on the relevant code.

As `interpret` does not perform any check for stack overflows, a call to
a Generator Function or Method or the processing of a resume operation
on a Generator will not identify those overflows, such that it will fail
to throw a `RangeError`, as it would be normal for a normal call,
instead ending up segfaulting down the call chain.

To avoid the issue, ensure that a check for stack overflows is performed
before a call to `interpret` in both `GeneratorFunction::virtualCall`
and `GeneratorObject::resume`, to ensure that an exception is thrown
when required.

The ordering of operations for `resume` was modified to allow for the
abrupt return due to the, now possible, stack overflow exception.

Both `return` and `throw` can produce a resume of the generator and both
of them throw an exception in doing so.

`return` throws an exception as an implementation detail.
An `emptyValue` is thrown, which will later be recognized by the
execution of a `Resume` instruction and cleaned away.
Failure to do the cleaning will produce an issue as an `emptyValue`
should not propagate through the system.

`throw` throws as part of its implementation as expected by the spec.

As both throws are performed before resume is called, the thrown value
would hide the `RangeError` exception that can be thrown due to a stack
overflow.

To allow for the stack overflow to be correctly recognized and its
thrown error not to be hidden, `resume` was modified to accept a third
optional parameter, representing a value to be thrown just before
calling `interpret` and after checking for a stack overflow.

Its call site in `method_next`, `method_return` and `method_throw` were
modified accordingly.

Some test-cases were added to `tst_qjsengine` to ensure that the
correct kind of exception is thrown when a stack overflow happens on the
execution of a Generator Function or Method or when resuming a Generator.

Pick-to: 6.7
Fixes: QTBUG-127049
Change-Id: I3656957097df3ccd5c6d70f86335bc40e229bb20
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit d408caa701448f8fe333894831c63b6bca74830a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>