VectorImage: Fix gradients applied to individual letters
When the gradient is in object mode, it will be adapted to the
bounding rect of the path it draws. However, since we are creating
paths for each letter individually when doing this for text, we
ended up applying the gradient to each letter's bounding box.

This patch allows the caller to override the path's bounding rect
with a different one, which in the case of text will be the
QTextBlock's bounding rect instead.

Fixes: QTBUG-126235
Change-Id: I65fa994c98d1e13154c2f5aa218ddf0353388a79
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
(cherry picked from commit c25bfd53425894ff12f3f883f3d6c211e9840973)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>