Tumbler: Prevent changing currentIndex when creating the view
The initial value of the QQuickTumbler was not the same as the value in
the QQuickListView when the Tumbler component was loaded. This meant
that the view was showing an incorrect item in the Tumbler. The reason
was that during the QQuickTumbler creation, the currentIndex would
change due to the QQuickTumblerView::createView() in the view
initialization.

To fix this issue, the QQuickTumbler now ignores changes to currentIndex
until the view is initialized. When the view has initialized and
QQuickListView's currentIndex is not the same as the QQuickTumbler's
currentIndex, it will be reset to the QQuickTumbler's currentIndex value
which is set in Tumbler by the user.

The existing test case has been updated to support the Tumbler without a
delegate to force the default view to be instantiated. The currentIndex
is set to a value outside the highlight range, and it will then be
checked that the currentIndex remains the same as the initial
currentIndex.

Fixes: QTBUG-127315
Pick-to: 6.7 6.5
Change-Id: I2603cfba228728d6fd3c464dce2e4eeac1365d75
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 6b440f5a0c62508405a5f0a0d757b0e83effffe8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>