Tumbler: Set view currentIndex when currentIndex is set on modelChanged
The Tumbler displayed an incorrect item when the currentIndex was set
during a model change. This issue occurred because setting currentIndex
in the onModelChanged handler was not considered. The bug happened when
the Tumbler was instantiated with the wrap property set to false.

To fix this issue, it should be verified that the view's currentIndex
matches the QQuickTumblerPrivate's currentIndex after endSetModel().
The currentIndexSetDuringModelChange property can help determine if the
currentIndex was set during a model change.

There were a couple of test cases that did not account for this
scenario. The test cases have been updated to be more general and to
check for this and other related cases.

Fixes: QTBUG-124921
Pick-to: 6.7 6.5
Change-Id: I82511cee8a5dbd6295fce6b3451aa205e977a8f5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit ca771d0a018138ede48f3b7bf6732a16650bd888)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>