DialogButtonBox: Fix docs for default alignment property
The default value for the alignment property is not undefined, as this
is an integer property. Correct this in the docs.
This is a resettable property and setting it to undefined is
allowed and will call the RESET function, however the returned value can not be undefined.

Pick-to: 6.5
Change-Id: Id9d2f74052fa4a85a8505378d41c990fd7b3283c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 3c0207c00cbf6a05c34c706df6eace9d8d411b03)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 84c3ca6090005882947e547910b7cddfb563fe7b)